fix: restore buildability, correct spectrogram query, harden CPPS (v4.9.19)
Follow-up to the assessment in dev/ASSESSMENT_2026-08-05.md. v4.9.18 did not
compile and did not load; this release fixes that, removes a numerically wrong
optimisation it introduced, and closes several correctness and documentation
gaps the assessment measured.

Build
-----
* batch_queries.cpp: correct MelderThread_PARALLELIZE/FOR/ENDFOR brace usage.
  The macros open their own scope; the extra braces broke compilation.
* Regenerate src/RcppExports.cpp, stale relative to calculate_cpps_ultra_cpp's
  signature, which caused "symbol not found in flat namespace" at load.

Correctness
-----------
* Spectrogram$get_power_at() now matches Praat. It did a nearest-cell lookup;
  Praat's "Get power at (time, frequency)" is Matrix_getValueAtXY, i.e. bilinear
  interpolation. Old results differed from Praat by up to ~24% at a point and
  disagreed with the same class's get_power_at_points(). Relative deviation vs
  Praat: 2.4e-01 -> 1.4e-08. Out-of-domain queries now return NA.
* Remove calculate_cpps_ultra(fused=). It reused the trend fitted on the raw
  cepstrum as the baseline for the peak measured on the smoothed,
  trend-subtracted cepstrum; Praat fits twice deliberately. Returned -47.169 dB
  where the correct value is 9.9205 dB, and was 3.1x slower.
* Formant$get_value_at_time() returns NA for non-finite time instead of raising.
* Fix get_pitch_strengths_at_times(), which errored with "object 'unit_code' not
  found" after the v4.9.18 unit-code refactor.
* to_formant_keepall/willems/sl() accept time_step = 0 again; Praat means "auto".

Faithfulness
------------
* Registry 10 -> 14 routines, all passing. The four v4.9.18 additions could not
  run: the CPPS oracle omitted its leading boolean argument, the SHS oracle used
  a time step Praat rejects, one row duplicated an existing routine at an
  unachievable tolerance, and the keep-all row called a nonexistent method. New
  coverage: harmonicity, spectrogram, jitter, MFCC.
* Tests no longer write into the source tree; reports go to tempdir() unless
  PLADDRR_FAITHFULNESS_OUTDIR is set.

Robustness (design principle 6)
-------------------------------
* Validators raise classed pladdrr_input_error conditions, so the typed-error
  contract holds for argument validation without with_pladdrr_errors().
* extract_part()/extract_parts_batch() emit a pladdrr_data_loss warning when the
  window runs past the signal. Praat zero-pads silently; the value is preserved
  but no longer unannounced. options(pladdrr.data_loss=).
* Reversed quefrency ranges error instead of returning a different measurement.
* fit_method = "robust slow" warns once per session: Praat's Theil-Sen fit
  samples randomly and is not reproducible (~0.8 dB spread; Praat itself
  occasionally returns ~1e290). Upstream defect, reproduced faithfully.

Performance
-----------
* SIMD batch bridges pass R's storage through as values.begin() - 1 instead of
  copying into a 1-based std::vector. vs base R on 1e6 doubles: mean 0.87x ->
  4.1x, sd 1.10x -> 2.8x, range 2.33x -> 15.4x.
* calculate_quantile_simd() selects the two needed order statistics with
  nth_element instead of fully sorting: 0.47x -> 1.8x vs base R.
* MelderThread_computeNumberOfThreads() mirrors upstream's platform divisors
  instead of hardcoding the macOS branch (oversubscribed Windows ~2x, Linux ~1.5x).
* Rejected and recorded: std::nth_element in getSlope_Siegel is bit-exact but
  25% slower than Praat's median-of-ninthers. ~94% of CPPS remains that fit.

Housekeeping
------------
* Delete six *_simd.cpp files (1,174 lines) with no referenced symbols.
* configure no longer computes a @XSIMD_FLAG@ that Makevars.in overrode; drop
  the dead FLAC_SRC/MP3_SRC naming a praat/ tree removed in v4.9.5.
* Correct the FLAC/MP3 documentation: read via the suggested av package, not
  natively; the decoder sources were dropped in the v4.9.5 CRAN slimming.
* AGENT_GUIDE.md 9,240 -> 4,233 lines; version history and the superseded SIMD
  archive moved verbatim into HISTORY.md. ARCHITECTURE.md and
  PERFORMANCE_INVENTORY.md updated with measured figures and the compile-time
  SIMD reality (stock builds reach SSE2 on x86_64, NEON on arm64 — never AVX2).
* Remove unverifiable speedup claims from reference docs and both performance
  vignettes; fix broken example fixture paths; make the CPPS examples runnable.
* Test harness: call unexported internals via pladdrr:::, look up S3 methods in
  the namespace. Suite 28 failures + 15 errors -> 5 failures + 0 errors.

+++ tests/testthat/test-spectrum-r6.R
@@ -90,8 +90,11 @@ test_that("Spectrum R6 get bin number from frequency works", {
   
   # Get bin for 1000 Hz
   bin <- spectrum$get_bin_from_frequency(frequency = 1000)
-  
-  expect_type(bin, "integer")
+
+  # Praat's "Get bin number from frequency" is Sampled_xToIndex, which returns a
+  # REAL (fractional) bin index, not an integer - 1 + (f - x1)/dx. Expecting an
+  # integer here was a test bug, not an implementation bug.
+  expect_type(bin, "double")
   expect_gt(bin, 0)
 })
